--- references/cuegen/testdata/valid/exported.cue
@@ -1,16 +1,0 @@
-package test
-
-A: {
-	field1: string
-	field3: string
-}
-B: {
-	field1: string
-	field2: {
-		field1: string
-		field3: {
-			field1: string
-			field3: string
-		}
-	}
-}
