--- references/cuegen/testdata/valid/kube.cue
@@ -1,17 +1,0 @@
-package test
-
-ResourceReturns: $returns: {
-	...
-}
-Any: {
-	a: {
-		...
-	}
-	b: {
-		...
-	}
-	c: [string]: string
-	d: {
-		...
-	}
-}
